Outpass-Management-System-core-python/requirements.txt
streamlit==1.38.0
supabase==2.3.1
python-dotenv==1.0.1
requests
pandas
